Restore highway processing & better separate API used by plugins
 ~ highway snippets now use the .highway.cpp ext, which the new system can detect instead of highway_ prefix
 * designed so that plugins only need minimal access to src_file_compiler_builder.h and none of the implementations behind it.
